Add basic mobile styles
Add a mobile stylesheet copied from https://github.com/niqjohnson/plans-styles/.
Right now, this is a very basic mobile stylesheet that all users will see for
screen sizes 600px and below.

This commit also updates $page->stylesheets to be an associative array so we
can add media queries to the stylesheet <link>s that depend on the kind
of stylesheet.